Refactor, BUGFIX: move version check logic to mysql files
This commit refactors the version check / flavor logic, moves
everything to the MySQL-related source files. With this, everything
MySQL specific is now in the correct file.

The flavor (pxc/ps/mysql) detection is also based on this new
version/flavor API, the command line parameter is not used. But it
is kept, as we need it for choosing between MySQL variants and
PostgreSQL variants.

There was also a bug where previously the MySQL version selection
logic was based on the version of the client library pstress was
compiled with, not on the actual version of the server it is
connected to.

This commit also fixes that, and now the version is selected
dynamically at runtime based on the actual server.